[refactor] : 컴포넌트 관련 로직 컴포넌트와 가깝게 이동
